Change the structure of test folder
